feat: 円グラフをPieWithStep構造に変更し同色青に統一
- PieWithStepコンポーネント作成（同色青ドーナツ+ステップ形式ランキングリスト）
- SimplePieChartも同色青に変更しカラードットを順位表示に
- CategoryPieChart・assignChartColors・CHART_COLOR_GRAYを削除
