+++ frontend/src/index.css
@@ -30,7 +30,6 @@
   --color-chart-6: #93c5fd;
   --color-chart-7: #bfdbfe;
   --color-chart-8: #dbeafe;
-  --color-chart-gray: #94a3b8;
 
   /* Chart - Scale (blue gradient: light → dark) */
   --color-chart-scale-1: #dbeafe;
